output consistency: move is_any_date_time_expression
